fix: P1-4 contract collection methods, P3-1 Char i32 consistency, and regression tests
P1-4 (contracts.xi): Implemented ContractIndex.none(), .is_sorted(), .contains_fn(), .contains_type(), .all_clauses(), .filter_nonempty() plus package-level convenience functions any_contracts(), fn_has_contracts(), type_has_invariants().

P3-1 (M1 Char mapping): Fixed Char from i8 to i32 in three locations — lib.rs part resolver (line 4607), infer_llvm_type_impl (line 4690), and types.rs dead-code mapping. Now consistent with the authoritative xiom_to_llvm_type mapping at lib.rs:911.

Regression tests: Added 6 new tests — 2 for Char i32 consistency, 4 for P1-4 contract collection methods.

+++ crates/xiom-codegen/src/lib.rs
@@ -4604,7 +4604,7 @@ let subst_elem = Self::substitute_type(t, elem, &type_map);
                 "Int" | "Bool" | "Int32" | "UInt32" | "UInt64" | "Int64" | "Int8" | "UInt8" | "Int16" | "UInt16" => "i64",
                 "Float64" => "double",
                 "Float32" => "float",
-                "Char" => "i8",
+                "Char" => "i32",
                 "Str" => "i8*",
                 _ => "i64", // default for unknown/custom types
             };
@@ -4687,7 +4687,7 @@ let subst_elem = Self::substitute_type(t, elem, &type_map);
             Expr::Int(_, _) | Expr::Bool(_, _) => "i64".to_string(),
             Expr::Float(_, _) => "double".to_string(),
             Expr::Str(_, _) => "i8*".to_string(),
-            Expr::Char(_, _) => "i8".to_string(),
+            Expr::Char(_, _) => "i32".to_string(),
             Expr::Ident(ident) => {
                 if let Some((_, llvm_ty)) = self.lookup_local(&ident.name) {
                     if llvm_ty == "double" { return "double".to_string(); }
